Refactor database code into separate file
